+++ src/components/views/dialogs/CreateRoomDialog.tsx
@@ -553,9 +553,6 @@ export default class CreateRoomDialog extends React.Component<IProps, IState> {
             createOpts.visibility = Visibility.Public;
             createOpts.preset = Preset.PublicChat;
             opts.guestAccess = false;
-            if (isVideoRoom) {
-                opts.encryption = this.state.isEncrypted;
-            }
             if (this.state.alias) {
                 const { alias } = this.state;
                 createOpts.room_alias_name = alias.substring(1, alias.indexOf(":"));
@@ -586,10 +583,7 @@ export default class CreateRoomDialog extends React.Component<IProps, IState> {
 
     public componentDidMount(): void {
         const cli = MatrixClientPeg.safeGet();
-        const isVideoRoom =
-            this.props.type === RoomType.ElementVideo || this.props.type === RoomType.UnstableCall;
-        const preset = isVideoRoom ? Preset.PublicChat : Preset.PrivateChat;
-        checkUserIsAllowedToChangeEncryption(cli, preset).then(({ allowChange, forcedValue }) =>
+        checkUserIsAllowedToChangeEncryption(cli, Preset.PrivateChat).then(({ allowChange, forcedValue }) =>
             this.setState((state) => ({
                 canChangeEncryption: allowChange,
                 isEncrypted: forcedValue ?? state.isEncrypted,
@@ -755,11 +749,9 @@ export default class CreateRoomDialog extends React.Component<IProps, IState> {
         }
 
         let e2eeSection: JSX.Element | undefined;
-        if (isVideoRoom || this.state.joinRule !== JoinRule.Public) {
+        if (!isVideoRoom && this.state.joinRule !== JoinRule.Public) {
             let microcopy: string;
-            if (isVideoRoom) {
-                microcopy = _t("create_room|encrypted_video_room_warning");
-            } else if (privateShouldBeEncrypted(MatrixClientPeg.safeGet())) {
+            if (privateShouldBeEncrypted(MatrixClientPeg.safeGet())) {
                 if (this.state.canChangeEncryption) {
                     microcopy = _t("create_room|encrypted_warning");
                 } else {