+++ src/components/views/location/SmartMarker.tsx
@@ -89,7 +89,6 @@ const SmartMarker: React.FC<SmartMarkerProps> = ({ id, map, geoUri, roomMember, 
                 roomMember={roomMember}
                 useMemberColor={useMemberColor}
                 tooltip={tooltip}
-                geoUri={geoUri}
             />
         </span>
     );