+++ src/stores/OwnBeaconStore.ts
@@ -55,8 +55,8 @@ export enum OwnBeaconStoreEvent {
     BeaconUpdateError = "BeaconUpdateError",
 }
 
-const MOVING_UPDATE_INTERVAL = 10000; // 10 seconds when moving
-const STATIC_UPDATE_INTERVAL = 10000; // 10 seconds when stationary
+const MOVING_UPDATE_INTERVAL = 5000;
+const STATIC_UPDATE_INTERVAL = 30000;
 
 const BAIL_AFTER_CONSECUTIVE_ERROR_COUNT = 2;