+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+Changes in [Unreleased] (2026-02-16)
+==================================================================================================
+## ✨ Features
+
+* Added encryption toggle for secure video conferencing.
+
+
 Changes in [1.11.91](https://github.com/element-hq/element-web/releases/tag/v1.11.91) (2025-01-28)
 ==================================================================================================
 ## ✨ Features